+++ NEWS.md
@@ -1,3 +1,5 @@
+# omock (development version)
+
 # omock 0.6.2
 
 # omock 0.6.1